requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.23
celery==3.1.23
Django==1.9.5
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
nltk==3.2.1
py-wikimarkup==1.0.2
python-dateutil==2.5.3
pytz==2016.3
redis==2.10.5
requests==2.9.1
six==1.10.0
Werkzeug==0.11.5
wikipedia==1.4.0
